some tests fail due to logical changes, will fix that in a follow-up

+++ includes/filebackend/FileBackendStore.php
@@ -89,6 +89,7 @@ abstract class FileBackendStore extends FileBackend {
 	 *   - content       : the raw file contents
 	 *   - dst           : destination storage path
 	 *   - overwrite     : overwrite any file that exists at the destination
+	 *   - disposition   : Content-Disposition header value for the destination
 	 *   - async         : Status will be returned immediately if supported.
 	 *                     If the status is OK, then its value field will be
 	 *                     set to a FileBackendStoreOpHandle object.
@@ -105,7 +106,9 @@ abstract class FileBackendStore extends FileBackend {
 		} else {
 			$status = $this->doCreateInternal( $params );
 			$this->clearCache( array( $params['dst'] ) );
-			$this->deleteFileCache( $params['dst'] ); // persistent cache
+			if ( !empty( $params['overwrite'] ) ) { // file possibly mutated
+				$this->deleteFileCache( $params['dst'] ); // persistent cache
+			}
 		}
 		wfProfileOut( __METHOD__ . '-' . $this->name );
 		wfProfileOut( __METHOD__ );
@@ -125,6 +128,7 @@ abstract class FileBackendStore extends FileBackend {
 	 *   - src           : source path on disk
 	 *   - dst           : destination storage path
 	 *   - overwrite     : overwrite any file that exists at the destination
+	 *   - disposition   : Content-Disposition header value for the destination
 	 *   - async         : Status will be returned immediately if supported.
 	 *                     If the status is OK, then its value field will be
 	 *                     set to a FileBackendStoreOpHandle object.
@@ -141,7 +145,9 @@ abstract class FileBackendStore extends FileBackend {
 		} else {
 			$status = $this->doStoreInternal( $params );
 			$this->clearCache( array( $params['dst'] ) );
-			$this->deleteFileCache( $params['dst'] ); // persistent cache
+			if ( !empty( $params['overwrite'] ) ) { // file possibly mutated
+				$this->deleteFileCache( $params['dst'] ); // persistent cache
+			}
 		}
 		wfProfileOut( __METHOD__ . '-' . $this->name );
 		wfProfileOut( __METHOD__ );
@@ -161,6 +167,7 @@ abstract class FileBackendStore extends FileBackend {
 	 *   - src           : source storage path
 	 *   - dst           : destination storage path
 	 *   - overwrite     : overwrite any file that exists at the destination
+	 *   - disposition   : Content-Disposition header value for the destination
 	 *   - async         : Status will be returned immediately if supported.
 	 *                     If the status is OK, then its value field will be
 	 *                     set to a FileBackendStoreOpHandle object.
@@ -173,7 +180,9 @@ abstract class FileBackendStore extends FileBackend {
 		wfProfileIn( __METHOD__ . '-' . $this->name );
 		$status = $this->doCopyInternal( $params );
 		$this->clearCache( array( $params['dst'] ) );
-		$this->deleteFileCache( $params['dst'] ); // persistent cache
+		if ( !empty( $params['overwrite'] ) ) { // file possibly mutated
+			$this->deleteFileCache( $params['dst'] ); // persistent cache
+		}
 		wfProfileOut( __METHOD__ . '-' . $this->name );
 		wfProfileOut( __METHOD__ );
 		return $status;
@@ -222,6 +231,7 @@ abstract class FileBackendStore extends FileBackend {
 	 *   - src           : source storage path
 	 *   - dst           : destination storage path
 	 *   - overwrite     : overwrite any file that exists at the destination
+	 *   - disposition   : Content-Disposition header value for the destination
 	 *   - async         : Status will be returned immediately if supported.
 	 *                     If the status is OK, then its value field will be
 	 *                     set to a FileBackendStoreOpHandle object.
@@ -235,7 +245,9 @@ abstract class FileBackendStore extends FileBackend {
 		$status = $this->doMoveInternal( $params );
 		$this->clearCache( array( $params['src'], $params['dst'] ) );
 		$this->deleteFileCache( $params['src'] ); // persistent cache
-		$this->deleteFileCache( $params['dst'] ); // persistent cache
+		if ( !empty( $params['overwrite'] ) ) { // file possibly mutated
+			$this->deleteFileCache( $params['dst'] ); // persistent cache
+		}
 		wfProfileOut( __METHOD__ . '-' . $this->name );
 		wfProfileOut( __METHOD__ );
 		return $status;
@@ -1460,7 +1472,9 @@ abstract class FileBackendStore extends FileBackend {
 	}
 
 	/**
-	 * Set the cached stat info for a file path
+	 * Set the cached stat info for a file path.
+	 * Negatives (404s) are not cached. By not caching negatives, we can skip cache
+	 * salting for the case when a file is created at a path were there was none before.
 	 *
 	 * @param $path string Storage path
 	 * @param $val mixed Information to cache
